Separate visual builders from model ones
