+++ src/adam/model/std_factories/std_model.py
@@ -6,9 +6,9 @@ import urdf_parser_py.urdf
 
 from adam.core.spatial_math import SpatialMath
 from adam.model.abc_factories import ModelFactory
-from adam.model.visuals import normalize_urdf_visual
 from adam.model.std_factories.std_joint import StdJoint
 from adam.model.std_factories.std_link import StdLink
+from adam.model.std_factories.urdf_visual import normalize_urdf_visual
 
 
 def urdf_remove_sensors_tags(xml_string):
